Add service tests to CI
