+++ hrm-domain/hrm-service/service-tests/ssm.ts
@@ -1,0 +1,37 @@
+/**
+ * Copyright (C) 2021-2023 Technology Matters
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see https://www.gnu.org/licenses/.
+ */
+
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { mockSsmParameters } from '@tech-matters/testing';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { Mockttp } from 'mockttp';
+
+export const mockEntitySnsParameters = async (
+  mockttp: Mockttp,
+  queueName: string,
+  topicName: string,
+) => {
+  await mockSsmParameters(mockttp, [
+    {
+      pathPattern: /.*\/queue-url-consumer$/,
+      valueGenerator: () => queueName,
+    },
+    {
+      pathPattern: /.*\/notifications-sns-topic-arn$/,
+      valueGenerator: () => topicName,
+    },
+  ]);
+};
